Implementing Encode end point
Changed return from using ToString to just passing in the char[] in the constructor since this gave me a string and not "System.Char[]" which I saw during debugging.
Added call to service in controller logic. Changed return type to give back a JSON object with the data we shortened.